@@ -3,14 +3,15 @@ import 'package:provider/provider.dart';
 
 import 'counter.dart';
 
-class CounterText extends StatelessWidget{
+class CounterText extends StatelessWidget {
+  const CounterText({super.key});
+
   @override
   Widget build(BuildContext context) {
     var counter = context.watch<Counter>();
     return Text(
       counter.count.toString(),
-      style: Theme.of(context).textTheme.headline4,
+      style: Theme.of(context).textTheme.headlineMedium,
     );
   }
-
 }
