@@ -6,7 +6,7 @@ import 'package:provider/provider.dart';
 import 'counter.dart';
 
 class MyHomePage extends StatelessWidget {
-  MyHomePage({Key? key}) : super(key: key);
+  const MyHomePage({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -15,16 +15,16 @@ class MyHomePage extends StatelessWidget {
       child: Scaffold(
         appBar: AppBar(
           title: Consumer<Counter>(
-            builder: (context, counter, _) => Text("My counter application: ${counter.count}"),
+            builder:
+                (context, counter, _) =>
+                    Text("My counter application: ${counter.count}"),
           ),
         ),
         body: Center(
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: <Widget>[
-              Text(
-                'You have pushed the button this many times:',
-              ),
+              Text('You have pushed the button this many times:'),
               CounterText(),
             ],
           ),
